Disentangle Kryo and AMQP classes.
